_index.tsx: Show events six weeks in the future (#92)
We meet every 2nd Thursday, so only showing meetups < 1 month away
means there will be times when no upcoming meetup is shown on the
homepage. Let's change this to 6 weeks, which also gives us some
wiggle room if we need to push to the 3rd week of the month, as we
have occasionally had to do.